Hookify slideout hash logic
